Moved classes to new folders
