+++ app/src/main/java/com/sage/cahekr/cakehrandroidapp/network/SWService.java
@@ -1,8 +1,9 @@
-package com.sage.cahekr.cakehrandroidapp;
+package com.sage.cahekr.cakehrandroidapp.network;
+
+import com.sage.cahekr.cakehrandroidapp.model.Film;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
-import retrofit2.http.Path;
 
 public interface SWService {
     @GET("films")
